backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
python-jose[cryptography]
pytest
pytest-asyncio
httpx
passlib[bcrypt]
